application/requirements.txt
attrs==21.4.0
blessed==1.19.1
boto3==1.26.19
botocore==1.29.19
chalice==1.27.3
click==8.1.3
cramjam==2.6.2
fsspec==2022.11.0
inquirer==2.10.1
jmespath==1.0.1
numpy==1.23.5
packaging==21.3
pandas==1.5.2
pyarrow==10.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-editor==1.0.4
pytz==2022.6
PyYAML==6.0
readchar==4.0.3
s3transfer==0.6.0
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
